Simplified the API
Now the Page and Book objects are more intelligent
about what they do with incomplete information, and
therefore the MutablePage no longer served a purpose.

The main backward-incompatibility is that now the Pages
don't add a Book-Title Suffix by themselves, but expect
whoever is creating the pages to add whatever suffix they
want.